+++ monitor/alert_me.py
@@ -1,6 +1,8 @@
 # -*-coding=utf-8-*-
 # 估价达到自己的设定值,发邮件通知, 每天2.45发邮件
 import fire
+import sys
+sys.path.append('..')
 from jsl_monitor import ReachTargetJSL
 from realtime_monitor_ts import ReachTarget
