+++ requirements.txt
@@ -7,3 +7,4 @@ twilio
 reids
 pymongo
 pymysql
+pypinyin
